Made generateD and E

+++ RSA/funcs/GeneratePrimeCandidate.swift
@@ -8,7 +8,7 @@
 func generatePrimeCandidate64() -> UInt64 {
         var candidate: UInt64
         repeat {
-                candidate = UInt64.random(in: 0 ..< UInt64.max)
+                candidate = UInt64(UInt8.random(in: 0 ..< UInt8.max))
                 candidate |= 1
         } while !RabinMillerTest(candidate)
         return candidate
